src-python/requirements.txt
easyocr==1.7.1
Pillow==10.0.1
PyGObject==3.46.0
PyGObject-stubs==2.9.0
Unidecode==1.3.7